Added the code to sync the custom settings with other players and fixed a bug
Added the patch to handle the custom menu options sync when one of them is change. This will make sure that all players have the same options. The bug that was fixed was because we were trying to access the jester code when the jester mod is not loaded. Currently the jester mod is not loaded as the custom menu options for both mods do not show up when both mods are used.